Iteration 7 - Extend other Commands from base BaseAccumulatedCommand
