+++ src/main/java/com/acme/dbo/txlog/command/ObjectCommand.java
@@ -1,62 +1,14 @@
 package com.acme.dbo.txlog.command;
 
-import com.acme.dbo.txlog.Controller;
-
-public class ObjectCommand extends BaseCommand {
+public class ObjectCommand extends BaseAccumulatedCommand {
     private String DECOR = "reference: ";
 
-    private Object currentValue;
-    private String accumulator;
-
     public ObjectCommand(Object message) {
-        currentValue = message;
-        accumulator = convertObjectToString(message);
+        super(message);
     }
 
     @Override
-    public Command accumulate(Controller controller, Command command) {
-        if (command instanceof ObjectCommand) {
-            ObjectCommand castedCommand = (ObjectCommand) command;
-            if (castedCommand.accumulator == null) {
-                this.accumulator = convertObjectToString(this.currentValue);
-            } else {
-                this.accumulator = castedCommand.accumulator + System.lineSeparator() + getDecoratedCurrentValue();
-            }
-        }
-        return this;
-    }
-
-    @Override
-    public String getCurrentValue() {
-        return convertObjectToString(currentValue);
-    }
-
-    @Override
-    public String getDecoratedState(int duplicationNum) {
-        String decoratedAccumulator;
-        if (duplicationNum >= 1) {
-            decoratedAccumulator = getDecoratedValue(accumulator) + " (x" + (duplicationNum + 1) + ")";
-        } else {
-            decoratedAccumulator = getDecoratedValue(accumulator);
-        }
-        return decoratedAccumulator;
-    }
-
-    @Override
-    public void flush() {
-        accumulator = null;
-    }
-
-    private String getDecoratedCurrentValue() {
-        return getDecoratedValue(convertObjectToString(currentValue));
-    }
-
-    private String getDecoratedValue(String object) {
+    protected String getDecoratedValue(String object) {
         return DECOR + object;
     }
-
-    private String convertObjectToString(Object character) {
-        return character.toString();
-    }
-
 }
